Add ability to renew auth
